Added history! - As a fragment (MediaChooserHistoryFragment) In the MediaChooser Activity - Uses the MediaBrowserService as a browser to get history - Service maintains history list. Adds to history when metadata changes
Changed "View tracks in an Album/Group" to use the MediaBrowserService as a browser - rather than using the MediaChooserTracksFragment, which browses media on the client side.

Note that MediaChooserTracksFragment is still used for displaying "All Songs" as it returns a huge list so we need pagination.

Also the "list of albums" and "list of artists" is still browsing locally on the client.